Install GI files, use SONAME-based package names, fixes
FIX: plain *.so symlink goes in the dev package.

Expand set of installed docs.

GObject-Introspection is part of the standard build if the libs are
present. Make sure they are for the puposes of repeatability, and add
a gir1.2-* package.

Addresses mypaint/libmypaint#28.

Name the binary libs package after the SONAME, current a "0" suffix.
https://www.debian.org/doc/debian-policy/ch-sharedlibs.html

Make clean conditional; don't want to have to reconfigure becore we can.

Exclude the .la file.
https://www.debian.org/doc/debian-policy/ch-files.html

Drop dh_installmenu since there is no menu file.